moving dummy classes to the unit test
